backend/requirements.txt
# FastAPI 백엔드 의존성 패키지 목록
# async SQLAlchemy, alembic, JWT, bcrypt 등 핵심 패키지
# pip install -r requirements.txt 로 설치

fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
strawberry-graphql[fastapi]>=0.243.0
